Move pdf_utils from browser/screenshot to components/screenshot
