+++ browser/ai_chat/BUILD.gn
@@ -139,6 +139,7 @@ static_library("ai_chat") {
     ]
   }
 
+  deps += [ "//brave//components/screenshot/content" ]
   if (enable_print_preview) {
     sources += [
       "print_preview_extraction_delegate_impl.cc",
