+++ browser/screenshot/BUILD.gn
@@ -9,8 +9,6 @@ assert(enable_print_preview)
 
 static_library("screenshot") {
   sources = [
-    "pdf_utils.cc",
-    "pdf_utils.h",
     "print_preview_extractor.cc",
     "print_preview_extractor.h",
     "print_preview_extractor_factory.h",
@@ -18,6 +16,7 @@ static_library("screenshot") {
 
   deps = [
     "//base",
+    "//brave//components/screenshot/content",
     "//chrome/browser/profiles",
     "//components/printing/common",
     "//components/printing/common:mojo_interfaces",
