+++ components/ai_chat/content/browser/BUILD.gn
@@ -44,6 +44,7 @@ static_library("browser") {
     "//brave/components/ai_chat/core/common",
     "//brave/components/ai_chat/core/common:common_impl",
     "//brave/components/ai_chat/core/common/mojom",
+    "//brave/components/screenshot/content",
     "//brave/components/api_request_helper",
     "//brave/components/brave_stats/browser",
     "//brave/components/constants",
