+++ components/screenshot/content/BUILD.gn
@@ -1,0 +1,15 @@
+# Copyright (c) 2026 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+static_library("content") {
+  sources = [
+    "pdf_utils.cc",
+    "pdf_utils.h",
+  ]
+
+  deps = [
+    "//content/public/browser",
+  ]
+}
